+++ rest-client-quickstart/src/main/java/org/csdg8/auth/dto/CreateTokenRequest.java
@@ -1,0 +1,11 @@
+package org.csdg8.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+public class CreateTokenRequest {
+    public String username;
+    public String password;
+}
