+++ rest-client-quickstart/src/main/java/org/csdg8/auth/dto/RefreshAccessTokenResponse.java
@@ -1,0 +1,10 @@
+package org.csdg8.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+public class RefreshAccessTokenResponse {
+    public String accessToken;
+}
